refactor: separate middleware tests
